+++ src/us/kbase/workspace/kbase/WorkspaceServerMethods.java
@@ -6,9 +6,12 @@ import static us.kbase.workspace.kbase.ArgUtils.chooseDate;
 import static us.kbase.workspace.kbase.ArgUtils.getGlobalWSPerm;
 import static us.kbase.workspace.kbase.ArgUtils.wsInfoToTuple;
 import static us.kbase.workspace.kbase.ArgUtils.processProvenance;
+import static us.kbase.workspace.kbase.ArgUtils.toObjectPaths;
+import static us.kbase.workspace.kbase.ArgUtils.translateObjectData;
 import static us.kbase.workspace.kbase.ArgUtils.longToBoolean;
 import static us.kbase.workspace.kbase.ArgUtils.longToInt;
 import static us.kbase.workspace.kbase.ArgUtils.objInfoToTuple;
+import static us.kbase.workspace.kbase.IdentifierUtils.processObjectSpecifications;
 import static us.kbase.workspace.kbase.IdentifierUtils.processWorkspaceIdentifier;
 import static us.kbase.workspace.kbase.KBasePermissions.translatePermission;
 
@@ -24,6 +27,8 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
+import com.fasterxml.jackson.core.JsonParseException;
+
 import us.kbase.auth.AuthException;
 import us.kbase.auth.AuthToken;
 import us.kbase.auth.ConfigurableAuthService;
@@ -34,10 +39,15 @@ import us.kbase.common.service.Tuple9;
 import us.kbase.typedobj.core.TypeDefId;
 import us.kbase.typedobj.exceptions.NoSuchPrivilegeException;
 import us.kbase.typedobj.exceptions.TypeStorageException;
+import us.kbase.typedobj.exceptions.TypedObjectExtractionException;
 import us.kbase.typedobj.exceptions.TypedObjectSchemaException;
 import us.kbase.typedobj.exceptions.TypedObjectValidationException;
 import us.kbase.typedobj.idref.IdReferenceHandlerSetFactory;
 import us.kbase.workspace.CreateWorkspaceParams;
+import us.kbase.workspace.GetObjectInfo3Params;
+import us.kbase.workspace.GetObjectInfo3Results;
+import us.kbase.workspace.GetObjects2Params;
+import us.kbase.workspace.GetObjects2Results;
 import us.kbase.workspace.GrantModuleOwnershipParams;
 import us.kbase.workspace.ListObjectsParams;
 import us.kbase.workspace.ListWorkspaceInfoParams;
@@ -51,6 +61,7 @@ import us.kbase.workspace.WorkspacePermissions;
 import us.kbase.workspace.database.DependencyStatus;
 import us.kbase.workspace.database.ListObjectsParameters;
 import us.kbase.workspace.database.ObjectIDNoWSNoVer;
+import us.kbase.workspace.database.ObjectIdentifier;
 import us.kbase.workspace.database.ObjectInformation;
 import us.kbase.workspace.database.Permission;
 import us.kbase.workspace.database.Provenance;
@@ -59,15 +70,19 @@ import us.kbase.workspace.database.User;
 import us.kbase.workspace.database.Workspace;
 import us.kbase.workspace.database.WorkspaceIdentifier;
 import us.kbase.workspace.database.WorkspaceInformation;
+import us.kbase.workspace.database.WorkspaceObjectData;
 import us.kbase.workspace.database.WorkspaceSaveObject;
 import us.kbase.workspace.database.WorkspaceUser;
 import us.kbase.workspace.database.WorkspaceUserMetadata;
 import us.kbase.workspace.database.WorkspaceUserMetadata.MetadataException;
 import us.kbase.workspace.database.exceptions.CorruptWorkspaceDBException;
+import us.kbase.workspace.database.exceptions.InaccessibleObjectException;
 import us.kbase.workspace.database.exceptions.NoSuchObjectException;
+import us.kbase.workspace.database.exceptions.NoSuchReferenceException;
 import us.kbase.workspace.database.exceptions.NoSuchWorkspaceException;
 import us.kbase.workspace.database.exceptions.PreExistingWorkspaceException;
 import us.kbase.workspace.database.exceptions.WorkspaceCommunicationException;
+import us.kbase.workspace.database.refsearch.ReferenceSearchMaximumSizeExceededException;
 import us.kbase.workspace.exceptions.WorkspaceAuthorizationException;
 
 public class WorkspaceServerMethods {
@@ -75,6 +90,8 @@ public class WorkspaceServerMethods {
 	final private Workspace ws;
 	final private Types types;
 	final private URL handleServiceUrl;
+	final private URL handleManagerUrl;
+	final private AuthToken handleManagerToken;
 	final private int maximumIDCount;
 	final private ConfigurableAuthService auth;
 	
@@ -82,6 +99,8 @@ public class WorkspaceServerMethods {
 			final Workspace ws,
 			final Types types,
 			final URL handleServiceUrl,
+			final URL handleManagerUrl,
+			final AuthToken handleMgrToken,
 			final int maximumIDCount,
 			final ConfigurableAuthService auth) {
 		this.ws = ws;
@@ -89,6 +108,8 @@ public class WorkspaceServerMethods {
 		this.handleServiceUrl = handleServiceUrl;
 		this.maximumIDCount = maximumIDCount;
 		this.auth = auth;
+		this.handleManagerUrl = handleManagerUrl;
+		this.handleManagerToken = handleMgrToken;
 	}
 	
 	public ConfigurableAuthService getAuth() {
@@ -367,6 +388,80 @@ public class WorkspaceServerMethods {
 		
 		final List<ObjectInformation> meta = ws.saveObjects(user, wsi, woc, fac); 
 		return objInfoToTuple(meta, true);
+	}
+	
+	/** Get object information.
+	 * @param params the information request parameters.
+	 * @param user the user making the request.
+	 * @param asAdmin whether the request should be run with administrator privileges.
+	 * @return the object information.
+	 * @throws WorkspaceCommunicationException if a communication error with the storage system
+	 * occurs.
+	 * @throws CorruptWorkspaceDBException if corrupt data is found in the storage system.
+	 * @throws InaccessibleObjectException if a requested object is inaccessible.
+	 * @throws NoSuchReferenceException if a reference in a reference path does not exist.
+	 * @throws NoSuchObjectException if a request object does not exist.
+	 * @throws ReferenceSearchMaximumSizeExceededException if a search for an object traverses to
+	 * many other objects.
+	 */
+	public GetObjectInfo3Results getObjectInformation(
+			final GetObjectInfo3Params params,
+			final WorkspaceUser user,
+			final boolean asAdmin)
+			throws WorkspaceCommunicationException, CorruptWorkspaceDBException,
+				InaccessibleObjectException, NoSuchReferenceException, NoSuchObjectException,
+				ReferenceSearchMaximumSizeExceededException {
+		checkAddlArgs(params.getAdditionalProperties(), params.getClass());
+		final List<ObjectIdentifier> loi = processObjectSpecifications(params.getObjects());
+		final List<ObjectInformation> infos = ws.getObjectInformation(user, loi,
+				longToBoolean(params.getIncludeMetadata()),
+				longToBoolean(params.getIgnoreErrors()),
+				asAdmin);
+		return new GetObjectInfo3Results().withInfos(objInfoToTuple(infos, true))
+				.withPaths(toObjectPaths(infos));
+	}
+	
+	/** Get objects.
+	 * @param params the object request parameters.
+	 * @param user the user making the request.
+	 * @param asAdmin whether the request should be run with administrator privileges.
+	 * @param resourcesToDelete a container into which resources that must be destroyed after
+	 * they're no longer needed can be placed.
+	 * @return the objects.
+	 * @throws WorkspaceCommunicationException if a communication error with the storage system
+	 * occurs.
+	 * @throws CorruptWorkspaceDBException if corrupt data is found in the storage system.
+	 * @throws InaccessibleObjectException if a requested object is inaccessible.
+	 * @throws NoSuchReferenceException if a reference in a reference path does not exist.
+	 * @throws NoSuchObjectException if a request object does not exist.
+	 * @throws ReferenceSearchMaximumSizeExceededException if a search for an object traverses to
+	 * many other objects.
+	 * @throws TypedObjectExtractionException if an error occurred extracting data from a typed
+	 * object.
+	 * @throws JsonParseException if an error occurs when attempting to get the object data. 
+	 * @throws IOException if an error occurs when attempting to get the object data.
+	 */
+	public GetObjects2Results getObjects(
+			final GetObjects2Params params,
+			final WorkspaceUser user,
+			final boolean asAdmin,
+			final ThreadLocal<List<WorkspaceObjectData>> resourcesToDelete)
+			throws CorruptWorkspaceDBException, WorkspaceCommunicationException,
+					InaccessibleObjectException, NoSuchReferenceException, NoSuchObjectException,
+					TypedObjectExtractionException, ReferenceSearchMaximumSizeExceededException,
+					JsonParseException, IOException {
+		// ugh, tried to see if we could code out the jpe and ioe but it's too much of a mess
+		// once you get into the byte array cache and JTS
+		checkAddlArgs(params.getAdditionalProperties(), GetObjects2Params.class);
+		final List<ObjectIdentifier> loi =
+				processObjectSpecifications(params.getObjects());
+		final boolean noData = longToBoolean(params.getNoData(), false);
+		final boolean ignoreErrors = longToBoolean(params.getIgnoreErrors(), false);
+		final List<WorkspaceObjectData> objects = ws.getObjects(
+				user, loi, noData, ignoreErrors, asAdmin);
+		resourcesToDelete.set(objects);
+		return new GetObjects2Results().withData(translateObjectData(
+				objects, user, handleManagerUrl, handleManagerToken, true));
 	}
 	
 	public void grantModuleOwnership(final GrantModuleOwnershipParams params,